adding fields to custom entities is working
