+++ packages/app-builder-backend/src/platform/index.ts
@@ -134,7 +134,7 @@ export const createMetadataSchema = async (schema: CreateSchemaRequest, accessTo
 }
 
 export const updateMetadataSchema = async (id: string, schema: UpdateSchemaRequest, accessToken: string) => {
-  const res = await fetch(`${platformApiUrl}/metadata/metadataSchema/${id}`, {
+  await fetch(`${platformApiUrl}/metadata/metadataSchema/${id}`, {
     method: 'PUT',
     headers: {
       Authorization: `Bearer ${accessToken}`,
@@ -144,9 +144,7 @@ export const updateMetadataSchema = async (id: string, schema: UpdateSchemaReque
     redirect: 'follow',
     body: JSON.stringify(schema),
   })
-  const parts = res.headers.get('location')?.split('/') || []
-  const newId = parts[parts.length - 1]
-  return getMetadataSchema(newId, accessToken)
+  return getMetadataSchema(id, accessToken)
 }
 
 const ENTITY_ID = 'dataType'
