+++ packages/app-builder/src/components/hooks/custom-entities/use-update-custom-entity.ts
@@ -1,11 +1,13 @@
 import { gql, useMutation } from '@apollo/client'
+import cloneDeep from 'clone-deep'
+import omitDeep from 'omit-deep'
 import { introspectionQuery } from '../use-introspection'
 import { CustomEntity, CustomEntityFragment } from './types'
 
 export const UpdateCustomEntityMutation = gql`
   ${CustomEntityFragment}
   mutation _updateCustomEntity($id: String!, $customEntity: _CustomEntityInput!) {
-    _updateCustomEntity(customEntity: $customEntity) {
+    _updateCustomEntity(id: $id, customEntity: $customEntity) {
       ...CustomEntityFragment
     }
   }
@@ -20,7 +22,7 @@ export const useUpdateCustomEntity = () => {
     updateCustomEntity: (id: string, customEntity: CustomEntity) =>
       updateCustomEntity({
         variables: {
-          customEntity,
+          customEntity: omitDeep(cloneDeep(customEntity), ['__typename']),
           id,
         },
         refetchQueries: [introspectionQuery],
