+++ packages/app-builder/src/components/ui/user/table.tsx
@@ -1,8 +1,7 @@
 import React from 'react'
-import { ToolbarItem, ToolbarItemType, ToolbarSection } from '../toolbar'
-import { useTypeList } from '@/components/hooks/objects/use-type-list'
-import Container from './container'
 import { useEditor, useNode } from '@craftjs/core'
+import { ToolbarItem, ToolbarItemType, ToolbarSection } from '../toolbar'
+import Container from './container'
 import { DestinationPage } from './link'
 import { TableProps, Table as ETable } from './ejectable/table'
 import { useSubObjects } from '@/components/hooks/objects/use-sub-objects'
@@ -12,8 +11,7 @@ import { getAvailableIntegrationsForArgs } from '@/core/desktop-integration'
 import { useObjectList } from '@/components/hooks/objects/use-object-list'
 import { useIntrospection } from '@/components/hooks/use-introspection'
 import { CreatePage } from './create-page'
-import { useCreateCustomEntity } from '@/components/hooks/custom-entities/use-create-custom-entity'
-import { Button } from '@reapit/elements'
+import { TypeList } from './type-list'
 
 const defaultProps = {}
 
@@ -57,41 +55,6 @@ export const IntegrationLanding = ({ typeName }: { typeName: string | undefined 
         <option value="">Select a page</option>
       </ToolbarItem>
     </ToolbarSection>
-  )
-}
-
-const TypeList = () => {
-  const { data, loading } = useTypeList()
-  const { createCustomEntity, loading: createLoading } = useCreateCustomEntity()
-
-  return (
-    <>
-      <ToolbarItem type={ToolbarItemType.Select} propKey="typeName" title="Object Type">
-        {(data || []).map((typeName) => (
-          <option key={typeName} value={typeName}>
-            {typeName}
-          </option>
-        ))}
-        <option value="" disabled>
-          {loading ? 'Loading...' : 'Select a Type'}
-        </option>
-      </ToolbarItem>
-      <Button
-        loading={createLoading}
-        onClick={() => {
-          const typeName = prompt('Enter the name of the new type')
-          if (typeName) {
-            createCustomEntity({
-              id: typeName,
-              name: typeName,
-              fields: [],
-            })
-          }
-        }}
-      >
-        Add New
-      </Button>
-    </>
   )
 }
